+++ src/DataTransferObjects/Results/Availability.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Iabduul7\ThemeParkAdapters\DataTransferObjects\Results;
+
+class Availability extends Result {}