+++ src/DataTransferObjects/Results/RatePriceSchedule.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Iabduul7\ThemeParkAdapters\DataTransferObjects\Results;
+
+class RatePriceSchedule extends Result
+{
+    /**
+     * @return array<string, mixed>
+     */
+    public function getPriceData(): array
+    {
+        return $this->getData();
+    }
+}